create a matching module.  refactor file matching logic.
