+++ core/signatures.go
@@ -4,10 +4,9 @@ import (
 	"crypto/sha1"
 	"fmt"
 	"github.com/codeEmitter/gitrob/common"
+	"github.com/codeEmitter/gitrob/matching"
 	"io"
-	"path/filepath"
 	"regexp"
-	"strings"
 )
 
 const (
@@ -15,31 +14,6 @@ const (
 	PartFilename  = "filename"
 	PartPath      = "path"
 )
-
-var skippableExtensions = []string{".jpg", ".jpeg", ".png", ".gif", ".bmp", ".tiff", ".tif", ".psd", ".xcf"}
-var skippablePathIndicators = []string{"node_modules/", "vendor/bundle", "vendor/cache"}
-
-type MatchFile struct {
-	Path      string
-	Filename  string
-	Extension string
-}
-
-func (f *MatchFile) IsSkippable() bool {
-	ext := strings.ToLower(f.Extension)
-	path := strings.ToLower(f.Path)
-	for _, skippableExt := range skippableExtensions {
-		if ext == skippableExt {
-			return true
-		}
-	}
-	for _, skippablePathIndicator := range skippablePathIndicators {
-		if strings.Contains(path, skippablePathIndicator) {
-			return true
-		}
-	}
-	return false
-}
 
 type Finding struct {
 	Id              string
@@ -88,7 +62,7 @@ func (f *Finding) Initialize(isGithubSession bool) {
 }
 
 type Signature interface {
-	Match(file MatchFile) bool
+	Match(file matching.MatchFile) bool
 	Description() string
 	Comment() string
 }
@@ -107,7 +81,7 @@ type PatternSignature struct {
 	comment     string
 }
 
-func (s SimpleSignature) Match(file MatchFile) bool {
+func (s SimpleSignature) Match(file matching.MatchFile) bool {
 	var haystack *string
 	switch s.part {
 	case PartPath:
@@ -131,7 +105,7 @@ func (s SimpleSignature) Comment() string {
 	return s.comment
 }
 
-func (s PatternSignature) Match(file MatchFile) bool {
+func (s PatternSignature) Match(file matching.MatchFile) bool {
 	var haystack *string
 	switch s.part {
 	case PartPath:
@@ -153,16 +127,6 @@ func (s PatternSignature) Description() string {
 
 func (s PatternSignature) Comment() string {
 	return s.comment
-}
-
-func NewMatchFile(path string) MatchFile {
-	_, filename := filepath.Split(path)
-	extension := filepath.Ext(path)
-	return MatchFile{
-		Path:      path,
-		Filename:  filename,
-		Extension: extension,
-	}
 }
 
 var Signatures = []Signature{
