+++ matching/filematching.go
@@ -1,0 +1,41 @@
+package matching
+
+import (
+	"path/filepath"
+	"strings"
+)
+
+type MatchFile struct {
+	Path      string
+	Filename  string
+	Extension string
+}
+
+var skippableExtensions = []string{".jpg", ".jpeg", ".png", ".gif", ".bmp", ".tiff", ".tif", ".psd", ".xcf"}
+var skippablePathIndicators = []string{"node_modules/", "vendor/bundle", "vendor/cache"}
+
+func (f *MatchFile) IsSkippable() bool {
+	ext := strings.ToLower(f.Extension)
+	path := strings.ToLower(f.Path)
+	for _, skippableExt := range skippableExtensions {
+		if ext == skippableExt {
+			return true
+		}
+	}
+	for _, skippablePathIndicator := range skippablePathIndicators {
+		if strings.Contains(path, skippablePathIndicator) {
+			return true
+		}
+	}
+	return false
+}
+
+func NewMatchFile(path string) MatchFile {
+	_, filename := filepath.Split(path)
+	extension := filepath.Ext(path)
+	return MatchFile{
+		Path:      path,
+		Filename:  filename,
+		Extension: extension,
+	}
+}
